core: Add partial recovery for allocation run corruption
The existing code was behaving badly in the case of allocation corruption
in that if corruption was found allocations after the run were not. This
change makes it so that we pick up with the first free allocation in the
run that is reachable via either a fast bin list for the corresponding
arena or a doubly linked list for that arena.

The fix is still suboptimal in that a long run of used allocations immediately
after the corruption will be lost, up to the point of the next free allocation.
I'll make a more thorough fix later.

In passing, this commit fixes a corner case where someone has overridden the
default maximum heap size at compile time, so that the check for whether to
print a rather scary set of messages about missing heaps is done correctly.