core: Improve the corruption checks a bit and add some recovery
Do a better job flagging fast bin errors and errors found in the doubly linked
free list.  Use the free lists as a way to recover when a write past the end
of an allocation corrupts an allocation run, so that we can at least recover
starting with the next free allocation.

In passing, look for allocations eagerly, because it seems best for the user
to find out about corruption immediately, as opposed to waiting for the first
command related to allocations.

I have started to add some test code for corruption issues.  For example,
UseAfterFree corrupts both the fast bin and doubly linked lists, but I haven't
added the corresponding expectedOutput tests, partly because IIRC the harness
doesn't yet have a way to expect particular error messages.  I think I will add
a test when I finish off the "taint" subsystem.