test: Add a simple test using symbols
This adds a very simple test using just two symbols, and a core that has the mangled names for both present in the core.
Clearly more such tests are needed, for example to handle cases where the mangled names are not present in the core.

In passing, fix a bug where if a class name looked like a hexadecimal representation of a number, the class name was
instead treated as a numeric pseudo-signature.