core: Add support for "freeoutgoing"
This implements a suggestion by Robin Kuzmin to allow the user to ask, for
a given allocation, what are the free allocations referenced by that
allocation.  To do this it was necessary to extend the graph so that it
keeps track of all references between allocations, not just references between
used allocations.  This means that existing iterators, such as Outgoing, that
are used for references to or from only used allocations, the check for
allocation status must be done in the iterator.

As examples of the new commands:
count freeoutgoing 6030b0   - counts the free allocations referenced by
                              6030b0
list freeoutgoing 6030b0   - lists the free allocations referenced by
                              6030b0
... show, summarize, enumerate