core: Handle truncated cores more cleanly
There were issues with handling truncated cores.  In particular, when the
"-t" flag was used to specify that chap should just check for truncation
and stop, chap was attempting to find allocations.  This was not needed
as part of the truncated check and interfered with the general goal that
the truncation check should be much faster than normal chap startup.

In passing I added a simple truncation test for the 64 bit case but
realistically some more tests should be added in this area (for example,
the 32 bit case).