core: Add one more reference constraint switch, and help for all.
Add a new reference constraint switch that allows imposing a minimum on
the number of outgoing references to free allocations.  This is related to
a request by Robin Kuzmin during C++Now 2017 to make it easy to find used
allocations that refer to free ones.  Unfortunately at present pretty much all
such references are false references, so the switch, /minfreeoutgoing, does not
yet have the value that it eventually will after DWARF awareness significantly
reduces the number of false edges.

Add help for all the 5 new switches (/minincoming, /maxincoming, /minoutgoing,
/maxoutgoing, /minfreeoutgoing).

Fix a bug in IsUnreferenced that I introduced when I started allowing looking
at references to and from free allocations, where I was failing to regard a
leaked allocation that had no incoming references from used allocations, but at
least one from a freed allocation, as leaked (which is correct) but not
unreferenced.