core: Support cores which have 2**32 or more references
Previously, cores with 2**32 or more references caused a crash.  After this
change, such cores work (as long as the allocation count is strictly less than
2**32).  The fix is suboptimal, as it adds 2*(8-4)*numAllocations bytes to the
memory used by chap in the case of a 64-bit core.

At some point we should make it so that if fewer than 2**32 edges are needed,
we use roughly the old amount of memory, and also so that cores with more than
2**32 allocations are supported.