core: Allow extending allocation sets
A common use case, when one is analyzing a core file to understand memory
growth, is to try to understand, given a set of allocations, what else
this holds in memory.  For example we might know that we have a collection
of objects that themselves have collections.  The exact impact of the outer
collection may be time consuming to assess if the inner collections have
variable sizes.  At some point in the future, DWARF-based type awareness
will allow removal of enough false edges that we can calculate
the impact of a certain set by calculating what would no longer be anchored
if the given set were gone.

Another use case is, given a particular set, for example of objects only of
a given type, to gather a small bit of extra information associated with
certain fields of those objects.  Again DWARF-based type awareness will give
us a quick way to know where the relevant fields.  An alternative now is
to use the python bindings of gdb to display the values of the fields, but
this is really slow for large sets.

However, given that DWARF-based type awareness is a way off, given that at
present I am prioritizing it behind support for multiple allocators, I have
provided a work-around to handle those use cases using a new "/extend" switch
that allows one to extend any set based on incoming or outgoing edges.
Multiple "/extend" switches can be provided on a single command line and
this is expressive enough to allow one to do either of the use cases above,
provided one is willing to figure out either relevant field offsets in advance
(perhaps using gdb to do this) or in some cases by specifying relevant type
names.

I have described this a bit more in the comments of an added test file
(test/expectedOutput/ELF64/LibcMalloc/HasContainersAndSymbols/run) and there
are several examples in the test.  At some point I will describe this in the
user guide.  I also need to describe the use of the /extend switch as part
of the help.