core: Limit redirect paths to 255 characters.
Use of "redirect on" causes the output of each command to go to a path
starting with the name of the core and followed by information about
the command.  The new "/extend" switch makes it more common that such
commands are long, which can result in names too large to be used on Linux.
For now the fix is to truncate long paths at 255 characters.