core: Add a default describer
There are still cases for which addresses are known in the core but the
describe command does not properly describe them.  Add a default describer
that at least mentions that the address is known and provides the permissions,
and that will give some information about the region and its mapping status
in case an explanation is requested.