core: Improve support for non-standard glibc maximum arena sizes.
It is possible to override the default maximum arena size for glibc.  Whether
it is advisable to override it is another issue, but this change makes the
logic in chap to detect and address this a bit more robust in that it will
now confirm any guess about the maximum heap size with checks that the found
ring of arenas is has consistently derived field offsets.