core: At support for thread cached allocations.
Recent versions of glibc support thread-local caches of freed allocations,
subject to a compile time switch and additional runtime configuration.
Although this change to chap is not an endorsement of the recent change for
thread local cache, it does support that feature, if present, by considering
allocations in the thread-local cache to be free and allowing a keyword called
"threadcached", so that one can use commands such as "count threadcached",
"list threadcached", "show threadcached"...   At some point in the future it
will probably be desirable that the "explain" command, when applied to a set
that contains at least one thread cached allocation, will mention the thread
number associated with that allocation.