core: Improve handling non-main arenas lost to truncation
This partially improves the case that a core file is so badly truncated that
none of the non-main arena headers are present in the core, and the main arena
header is not found, but heaps actually were found.  These changes will make
it at least possible to analyze such truncated cores for growth, but, as in
the previously handled truncation cases, some allocations will not be reported,
some allocations may wrongly be reported as used rather than free, and some
allocations may falsely be reported as leaked.