core: Add a framework for lightweight unsigned object recognition.
For unsigned allocations, the long term goal is to identify the type of an
allocation by using symbol information.  This doesn't do that, because DWARF
awareness is not in place, but it allows one to create a PatternRecognizer,
that will, based on the contents of the allocation and possibly on some
incoming edges, make a partial guess about the type.

At present, this guess will be shown in the output of the "describe" and
explain commands.  At some point in the near future it will possible to use
a pattern name in places where one would use a signature.  At present the
only pattern supported is for the obsolete COW string body but I plan to
add support for other patterns in the near future.