core: Fix crash when thread cache disabled and describing free allocation
In the case that there were no thread cached allocations (often due to
using an older release of libc malloc prior to thread caching, but
potentially due to gathering a core extremely early in the life of a
process) calling IsThreadCached on an allocation index would cause a
crash.  This crash was not seen until now because IsThreadCached is called
only in rather rare cases, such as an attempt to describe or explain a
free allocation.