core: Get rid of lazy calculation of the graph
The two most common use cases (understanding growth and finding leaks)
require the graph anyways.  Checking for corruption or scanning for particular
patterns in memory do not, but the gain in startup time for those 2 less common
cases that one could get by lazy graph allocation are outweighed by the
relative ease of testing when the graph is known to exist (or not be possible
to create) as of the first prompt.