core: Identify symbols for anchors if available
Prior to this, the user could arrange, by doing "source <corefilename>.symreqs"
from gdb on the given core, to create <corefilename>.symdefs.  The use of this
was for cases where mangled class names were not present in the core.  This
change causes the addresses of any anchors to be included in the .symreqs as
well, so that when the .symdefs is created it will give symbols for those
anchors.  At present, the resulting symbol information is only used in the
output of the explain command but it will eventually be incorporated into other
commands.