core: Allow finding modules when the ELF headers are not mapped
The original algorithm for finding modules depended on the ELF headers being
mapped.  This was limited in that it is not uncommon for the ELF headers not
to be mapped.  The added algorithm does not have that limitation.  There are
still some issues in that there are a few cases where one or more module
boundaries are not calculated correctly.  The name of the main program is also
still not reported correctly.  In passing, fix a few cases where exceptions
were happening in fairly common code paths and so slowing down the initial
calculations.