core: Allow using PT_NOTE of type 0x46494c45 to find modules
For cores generated by relatively current versions of gdb there is a PT_NOTE
section of type 0x46494c45 (ASCII for FILE but as bytes in little-endian
order ASCII for ELIF) that makes it possible to identify which binaries were
used by the program and which address ranges were used by those binaries.  This
change supports using that type of PT_NOTE, adds a command "list modules" that
allows the modules and their address ranges to be listed, and adds a test
involving a core that has the given note type.