core: Fix the case where a program header has non-zero mismatching sizes.
Some program headers in the core have non-zero sizes for the file region
size and the memory region size where the file region size is smaller than
the memory region size.  In such a case it is correct to treat it as equivalent
to saying that a prefix of the specified region is mapped starting at the
given offset in the file and that the remainder of the region is not mapped in
the file but has known permissions.