core: Allow use of %<pattern-name> in place of a signature
The general notion of pattern recognizers has existed for a bit but to date
use of it has been limited to the "describe" or "explain" command, where
allocations conforming to a particular declared pattern have been identified
as such.  With this change, everywhere a signature can be used to restrict
a set, a pattern name, preceded with "%", can be used instead.  So for example:

list used %COWStringBody # All string bodies.
list used /minincoming %COWStringBody=1 # Things that reference string bodies.
list used %COWStringBody / extend %COWStringBody<-   # bodies and referrers

This will become more valuable as more recognizers are added.