core: Keep track of the status of signature names
At present there are already signature names can become known; they can
be found from mangled type names in the core, if present, or they can be read
from the .symdefs file generated by gdb.  The set of choices will expand soon
but in any case it is already useful to understand, for a given core, whether
it is particularly useful to generate the .symdefs file or not.  To this end,
there is a new subcommand called "summarize signatures".  In passing,
symbol unmangling was made more robust to reduce the dependency on gdb for
signature names, two tests were added, and the comments in some run scripts
were cleaned up.