core: Allow class and struct names to be extracted from unstripped binaries
Prior to this change, the only way to obtain class and struct names was to
follow from the vtable address to the type info to the mangled type name in
the core or to ask gdb, via "info sym" statements in the .symreqs file, to
show the class or struct associated with a given vtable pointer.  This meant
that there was no automated way in chap to determine the names if the mangled
names weren't present in the core and gdb was unable to open the core or didn't
have proper symbol information.

This change makes it so that chap, if the executable and libraries are present
at the same path they were on the server where the core was created, can
get the mangled type names from the binaries.  This is particularly useful if
the core was truncated because it got huge (and so gdb cannot open it) or if
the symbols are not available.  Particularly in the case of unexpected process
growth, the process could grow too large for the core to be stored but it
may still be possible to use chap to understand the growth from a partial core.