core: Add a pattern recognizer for std::vector bodies
At a pattern recognizer for std::vector bodies.  As with any pattern, this
means that when the describe command or explain command are used, if the
pattern is recognized for a given allocation, the description or explanation
is adjusted accordingly.  Also, as with any pattern, it can be used in any
command in place of a signature.

In passing, update some existing tests slightly to test the new pattern,
update USERGUIDE.md to mention it, and make some minor unrelated fixes for
some corner cases related to file truncation and to the case where there is
not enough memory where chap is running to handle the given core.