core: Add LongString pattern recognizer
The long string pattern recognizer recognizes the case where an allocation is
being used to hold the contents of an std::string that is at least 32
characters long.  As with any pattern name, the pattern name, prefixed by a '%'
can be used anywhere a signature can be used.

For example, the following describes all used allocations that match the
given pattern in the core:
describe used %LongString

In passing, update the documentation and add some simple tests.