core: Recover better from an overflowed main arena allocation
If the size/status word following an allocation gets clobbered, this
interferes with detection of subsequent allocations in the same run.  This is
true for all arenas, but the mitigation of this case for the main arena is a
bit weaker than the mitigation for the main arena, based on the fact that we
cannot count on knowing the boundaries of the runs, as we can in the case
of non-main arenas (because they have heaps).  This improves the main arena
case, which was failing to keep even some allocations before the overflowed
allocation.  There are still issues here but in practice this sort of overflow
seems less common than other forms of heap corruption.

In passing make some minor cosmetic fixes.