core: Fix a bug in the calculation of static and stack anchor points
This was quite a serious bug, where a free allocation could be marked as
a static anchor point or a stack anchor point.  In the case of static anchor
points this likely was happening somewhat rarely because it would require
statically allocated memory not used for administration of dynamically
allocated memory to point to free allocations.  The bug was exposed on a
core for a java program, because ranges used for java objects are allocated
outside of libc malloc and regions not understood to be related to threads or
allocations are currently considered to be statically allocated.

In passing, add a minor performance optimization to the explain command.  In
the case that an allocation is not anchored in a particular way (for example,
via the stack or by registers or statically) don't attempt to explain how it
might be anchored in that way.