core: Eliminate a scary but useless message related to finding modules
For cores generated with a relatively old version of gdb, chap cannot
simply use the entry in the PT_NOTE section to determine which shared libraries
were loaded and where.  There was an error message for one case where it was
attempting to figure out the module name for a region in one of these old cores
but failed.  The message was not particularly informative but a bit scary
looking.