core: Support "describe arenas"
Libc malloc has a notion of arenas, which roughly partition the memory
used and free allocations into non-overlapping sets, as a way of reducing
lock contention for thread allocations.  The "describe arenas" command helps
one understand the case where this feature of libc malloc drives up the
free allocation byte count to the point that it dominates the process size.