core: Add commands to get a high level understanding of virtual address use
For some processes, the most interesting information about address use is
outside of what one can find out about used allocations, or free
allocations or about stacks.  For example, if one wants to understand total
memory usage, the inaccessible regions used for heap tails (used to reserve
the part of each heap that has not yet been made accessible) or for module
alignment regions, or for stack guards all may have a significant impact
on the total memory usage.  In addition, a bug in gdb can make this a bit
more confusing because gdb sometimes marks such inaccessible regions as
read-only rather than as inaccessible.

To make things a bit more clear, chap now has commands to count or list or
describe regions associated with a particular set of permissions.
For example, "count inaccessible" will tell you how many inaccessible ranges
are associated with the process or "describe inaccessible" will list those
ranges and also provide a course indication of how each range is used.

In the particular case that the core was made by a buggy version of gdb
that marks the inaccessible regions as read-only, using "describe readonly"
will at least make it possible to see for cases where chap has detected that
the given regions were used for heap tails or module alignment regions or
stack guards.

Use a core compressed by bzip2 rather than a plain core because in this
particular case bzip2 shrinks the core by a factor of roughly 300.