doc: Fill in the sections on references in USERGUIDE.md
Now it gives at least sparse explanations of the difference between a real
reference and a false reference.  It might benefit from some sample code and
sample usage of chap.

The section on supported memory allocators is updated to reflect the fact
that other languages than C and C++, for example Python and Java, often
involve at least some use of libc malloc, for example due to use of native
methods.