test: Update the framework to delete files created in previous test runs
The script in test/expectedOutput/driver was leaving files created in
previous runs and so not distinguishing in the final recursive diff
between files created in the current run and files created in the previous
run.  In passing correct two .symreqs files to be what we currently expect
chap to create.  These were not detected under the old test driver because
chap currently doesn't bother to create the .symreqs file if it already
exists.