test: Fix bug just introduced in test/expectedOutput/driver
With my previous change, a change in the run script in the source
would cause the run script to be deleted in the test working directory.
Note that even after this change, changes to other inputs (such as core
files) will cause them to get deleted in the working directory but for
the moment this does not seem to be a problem because one should really
add a new test if new core files are used.  At some point a better fix
would be to have some record of what all the inputs are and to clean up
everything else before the test.