code: Improve the corruption check for the doubly linked free list
It wasn't checking that the target of the forward link, in the case that
the target was not the free list header, was at a suitable offset from the
start of the allocation.  Adding this check improves the logging from
complaining about the target, which might be totally well formed but not belong
on the list, to complaining about the list node with the invalid forward link.