core: Partially allow writable vtables
Some linkers produce shared libraries or executables that result in writable
vtables.  This is a mild security violation in that someone can overwrite those
vtables.  Another side effect is that it prevents chap from detecting those
vtables.

With this change, in the specific case that the vtable is (wrongly) writable
but the mangled type name is present in the core, support deriving the
signature name from that information.  At some time in the near future it might
be good to support a similar case where the mangled type name is only in the
binary but it is not clear how we'd ever support the case where the mangled
type name is not in either place.  Also, at present the code is fairly verbose
in the case of such signatures and we might want it configurable whether we
simply don't recognize writable vtables (as happened before this change) or
whether we recognize them and warn (as the code does after this change) or
whether we accept such vtables silently.