core: Detect the case that a process was killed during core generation
If a process gets killed while the core is being generated, the result is
often that the core is full sized but many runs of pages are filled with 0.
This detects that and puts out a reasonably clear message so that the user will
know that a better core is needed.  The error lines look like this:

Apparently this core file was not completely filled in.
Probably the process was killed while the core was being generated.
As a result any commands related to allocations will be very inaccurate.