core: Add a pattern recognizer for instances of PyDictKeysObject for python 3.5
For Python, most of the allocation is done outside of libc, but instances of
PyDictKeysObject are allocated via libc malloc().  The new pattern
%PyDictKeysObject can recognize these instances, and when the "describe"
command is applied to a set of allocations containing at least one such
instance, any key:value pairs for which both the key and value are strings
are displayed in the output of the command.

This is nothing close to real support for python, but can be quite useful
depending on how heavily dict is used.