core: For the /extend switch allow extensions constrained to leaked referees
In analyzing the impact of a particular memory leak it can be helpful to
understand which allocations are leaked as a result of other allocations
being leaked.  To do this, one wants to be able to see which allocations
reached via outgoing references from a leaked allocation are also leaked.
To support this, the /extend switch also supports "~>" as an alternative to
"->", where "~>" constrains the extensions to the set to be leaked objects.

For example, to find out the impact of leaked instances of SSL_CTX, one might
use one of the following:
  summarize leaked %SSL_CTX /extend ~>
  count leaked %SSL_CTX /extend ~>