core: Make the describe command cover more sorts of addresses
This makes the describe command cover various sorts of libc malloc overhead
as well as any known sorts of inaccessible regions.  In passing, it removes
the assumption that the process image was subject to change, because any plans
to support files that hold multiple process images were dropped, and merges
the functionality of the VirtualMemoryPartition with
PermissionsConstrainedMappings, to avoid inconsistences between them and does
various sorts of cosmetic cleanup.  The "describe p" commands, where p is a
permissions constraint, such as "writable", now are more detailed, in ways that
are more consistent with the describers for individual addresses.  Tests are
also added for the new functionality.