core: Make the COWStringBody pattern recognizer a bit more accurate
Look more specifically at all the incoming references from other allocations
and from anchors, counting references only to the start of the c-string.

In passing, don't warn to standard error if an allocation matches multiple
patterns because this is already clear from the output and the added warnings
coming to the terminal can be jarring when redirect is on.