code: Add support for %DoublyLinkedListNode pattern
This adds a pattern recognizer for allocations that are doubly linked using
pointers at the start of the allocation, such as nodes used in an std::list.
In the case of a simple pattern match or the describe command, the recognizer
just looks at the given allocation and where it points.  In the case of the
explain command, the recogizer also attempts to identify the list header.