core: Correct handling for writable bytes after the top of a non-main arena
It can happen that in the last heap for a non main arena just after the end
of the top allocation (which is a free allocation) there are more writable
bytes.  This can happen if the _top entry is large and it gets shrunk by some
number of pages at the end, without actually making those pages inaccessible.
It is best to count this toward the free memory, because libc malloc still
owns that memory and could use it again without actually changing any
inaccessible pages to be RW.

This change counts that last region toward the size of the top allocation for
the arena, so that those bytes are counted toward the free allocations.  The
implementation of "describe arenas" also had to be adjusted accordingly.