core: Change ", " to "," in unmangled type names from the .symdefs file
Changing ", " to "," in unmangled type names still leaves them reasonably
readable but allows type names that have no embedded blanks other than blanks
after "," to be used as signatures.  This is more interesting for chap scripts
than manual use, as in the manual case it is pretty easy to use the
coresponding vtable address but the vtable address is subject to ASLR and
isn't really known until the process runs.  There are still some cases where
embedded blanks might happen, such as when a type name contains something like
"unsigned long" but these become less clear with stripping blanks so if we
want to allow entering such names symbolically a better strategy will be to
allow escaping the blanks.