core: Allow the user to specify the suffix in the case of output redirection
Normally, if the user has enabled output redirection, via "redirect on", chap
will calculate the path of the output for each chap command as the core file
name followed by some string derived from the chap command.  At times,
particularly when the /extend switch is used, the resulting file names may
be quite long.  This allows the user to specify the suffix to be appended to
the core name via "/redirectSuffix someSuffixHere".