core: Add support for recognizing the "map" of each deque
A deque has an outer allocation, which it calls a map, that points to
other allocations that actually hold the elements.  The pattern %DequeMap
recognizes this outer map.  There will be a separate pattern to recognize
the blocks of elements.

In passing, this fixes a false recognition of one of the element blocks for
a deque as being a vector body, specifically in the case that the _M_cur for
the _M_start or _M_finish of a deque matches the corresponding _M_first.