core: Weaken accuracy of %UnorderedMapOrSetNodes to improve performance
There was a tagging phase intended to catch the narrow case where a maximum
load factor larger than 1 was used and the number of buckets was 1.  This
phase was too expensive and requires modification.  As a short term fix,
disable that second phase calculation at the expense of missing a small
number of allocations used for nodes for unordered maps or unordered sets.