core: Pre-tag %COWStringBody pattern
This is part of the effort to pre-tag all the patterns so that it will be
reasonable to include pattern information in summaries.  As part of the
current change, logic to pre-check capacities and bucket counts and deque
map sizes agains the allocation size now relies on the allocation finder
to provide a rough lower bound of the memory request size that was likely
associated with each allocation.