core: Force that all patterns must be pre-tagged.
This is so that it continues to be the case that, for efficiency reasons, all
allocations that match a pattern are pre-tagged.  For this reason, all the
Recognizer classes ...Recognizer.h are renamed as ...Describer.h, and these
classes no longer actually check for the match but simply assume that this
has been checked before the Describe function is called.