doc: Update USERGUIDE.md with respect to "unrecognized" allocations
Document nsigned allocations that also don't match any pattern are considered
to be "unrecognized" and can be specified by "?" anywhere one could use "-" to
specify "unsigned".  Document that the "summarize" command, when applied to an
allocation or a set of allocations, summarizes individual patterns and now
summarizes "unrecognized" allocations rather than "unsigned" allocations.