core: Improve performance when modules can be found based on PT_NOTE entries
If the modules can be found easily via the PT_NOTE entries, it is better in the
case where libc malloc was used in a single threaded process to find the main
arena by scanning the libc module than to scan all unclaimed regions.

Similarly, if the modules can be found easily it is better to figure out early
where the python arenas are located, because it is cheap to do so, and claiming
those regions can make other scans cheaper.