core: Allow missing signatures if /allowMissingSignatures is set to "true"
The way signatures are recognized is first to notice that an allocation
has something that resembles a signature (most commonly, a vtable pointer)
then to figure out the name associated with that signature, possibly with
the help of gdb.  If no allocations have the given signature it will never
become known, which is problematic if one is writing a script and doesn't
know whether, at the time the script will run, the core will contain at
least one allocation for every signature used in the script.  Now someone
can say in advance, in the context of a specific command, that missing
signatures are OK by specifying "/allowMissingSignatures true" in the
command.

The default is still not to allow them, to reduce the chance that an
unnoticed typo will cause false conclusions to be reached.