core: Add partial support for python cores.
For Python (the cpython variant) 2.x and 3.x allocation boundaries are now
recognized and leak analysis appears to work.  Many of the types are
identified correctly but this is reflected only in the output for the
"describe" or "explain" commands and not in others such as "list" or
"so".  Also, there is no way at present to constrain members of a set
to match a particular python type such as "dict" or "tuple".

The reason this is not yet implemented is that python types fit
neither into the current notion of signatures, where a pointer-sized
object at the start of an allocation suffice to identify it or
patterns, where the set of known patterns is independent of any particular
core.  The likely fix here is to introduce the notion of a subpattern,
where the set of subpatterns for a given pattern depends both on that
pattern and possibly on the core.

I am not sure exactly when I will get back to this but should point out
that one can do memory growth analysis or memory leak analysis with the
current level of support.  If you are curious, try "describe used" on
a python core.