core: Allow a set to be derived from multiple commands
Most of the commands in chap operate on sets of allocations.  At times one
may be interested in more complex sets that the basic qualifiers and
extension operators support.  This commit adds support for deriving a set
by starting with one set of allocations and subtracting one or more other
sets.

To initialize the derived set based on some calculated set S, add the following
to any command that calculates S:
  /setOperation assign
To subtract some calculated set S' from the derived set, add the following to
any command that calculates S':
  /setOperation subtract

The derived set is then referred to as "derived".