code: Support python 3.6, 3.7 and 3.8
Note that python support is not actually complete, but these changes make chap
work for python 3.6, 3.7 and 3.8 about as well as it was already working for
python 3.5.

To get a sense of the current level of support, try "describe used" on a
core from one of those releases of chap.