core: Restrict modules checked for missing writable ranges
Some of the modules being checked where not shared libraries, causing a bit
too much chatter with warnings, because some of the things detected as
modules, especially for java programs for which we are checking for leaks
of native allocations, could not reasonably be expected to have writable
sections.  The current fix, to limit checking to things for which the path
looks like a .so is not quite right, in that it won't check the main
executable, but it is better than having no warnings at all.