core: Relax assumptions about writable regions for python library
There were assumptions that the pointer to the python arena structure
array was in the same memory as the static declaration of the python
type "type" and in the same memory as the headers for the python
garbage collection lists.  It turns out that neither assumption is safe,
so now the code just figures that all those things are somewhere in the
writable memory associated with the python library.