core: Strip blanks from mangled type names
This is so that any signatures including ones that might otherwise have
embedded blanks, such as "Foo<unsigned long>" can be used as tokens,
which do not allow embedded blanks.  So the above type would show as
Foo<unsignedlong>.  This may cause minor readability issues in any
signatures used in output, and would cause a problem if someone defined
a class/struct name as something like "unsignedlong", but overall it
seems better particularly for scripting to allow the use of any
signature by name, as opposed to having to resort to a numeric signature
which might vary even between runs of the process.