core: Support /geometricSample to limit the base visited allocation set
"/geometricSample <base-in-decimal>" when applied to any command that applies
to a set of allocations will cause only entries 1, b, b**2, b**3...
(relative to 1 as the first entry) to be visited.

For example:

   describe used /sample 2
   enumerated used Foo /sample 4

This is an alternative to redirecting the output when the set is large and
you are interested only in a sample and not the whole set.