code: Start to add code to support C++ inheritance
One cannot actually do anything useful with this yet but the various
type_info instances are recognized.  The goal is enventually to be
able to use a base C++ class name or struct name anywhere one might
be able to use the corresponding derived name.

In passing, move most of the C++-specific code together and fix a
bug in identifying the favored incoming reference of the root
%MapOrSetNode and add a few related tests.