core: Provide partial support for tcmalloc, golang and annotation
The support for cores from processes that use tcmalloc is missing support for
slab-allocated objects, and has not been well tested.  In particularly, the only tests
comitted here are for the gperftools/gperftools variant as opposed to the google/tcmalloc
variant.  The documentation has also not yet been updated to reflect this support.

The support for cores from processes that use golang finds many allocations but definitely
is not reliable for leak checking.  No testing has been committed yet and the documentation
does not yet reflect this support.

Support for annotation is intended to support annotating memory in allocations and
eventually memory in general.  The only kind of annotation supported so far is for
SSO strings.

One reason so many files were changed in this commit is that the copyrights all required
modification.

In passing, some files were changed to use some c++17 functionality.